ref #107: revert CMSModulePageTree to previous state (all changes with jstree are in the new Model NavigationTree)
